+++ src/rx-query.ts
@@ -1,6 +1,5 @@
 import {
     BehaviorSubject,
-    firstValueFrom,
     Observable,
     merge
 } from 'rxjs';
@@ -113,7 +112,6 @@ export class RxQueryBase<
     }
     get $(): Observable<RxQueryResult> {
         if (!this._$) {
-
             const results$ = this.collection.$.pipe(
                 /**
                  * Performance shortcut.
@@ -145,19 +143,7 @@ export class RxQueryBase<
                  * depending on query type
                  */
                 map((result) => {
-                    const useResult = ensureNotFalsy(result);
-                    if (this.op === 'count') {
-                        return useResult.count;
-                    } else if (this.op === 'findOne') {
-                        // findOne()-queries emit RxDocument or null
-                        return useResult.documents.length === 0 ? null : useResult.documents[0];
-                    } else if (this.op === 'findByIds') {
-                        return useResult.docsMap;
-                    } else {
-                        // find()-queries emit RxDocument[]
-                        // Flat copy the array so it won't matter if the user modifies it.
-                        return useResult.documents.slice(0);
-                    }
+                    return ensureNotFalsy(result).getValue();
                 })
             );
 
@@ -189,6 +175,7 @@ export class RxQueryBase<
 
     // time stamps on when the last full exec over the database has run
     // used to properly handle events that happen while the find-query is running
+    // TODO do we still need these properties?
     public _lastExecStart: number = 0;
     public _lastExecEnd: number = 0;
 
@@ -214,7 +201,7 @@ export class RxQueryBase<
     _setResultData(newResultData: RxDocumentData<RxDocType>[] | number | Map<string, RxDocumentData<RxDocType>>): void {
         if (typeof newResultData === 'number') {
             this._result = new RxQuerySingleResult<RxDocType>(
-                this.collection,
+                this as any,
                 [],
                 newResultData
             );
@@ -224,7 +211,7 @@ export class RxQueryBase<
         }
 
         const newQueryResult = new RxQuerySingleResult<RxDocType>(
-            this.collection,
+            this as any,
             newResultData,
             newResultData.length
         );
@@ -295,7 +282,7 @@ export class RxQueryBase<
      */
     public exec(throwIfMissing: true): Promise<RxDocument<RxDocType, OrmMethods, Reactivity>>;
     public exec(): Promise<RxQueryResult>;
-    public exec(throwIfMissing?: boolean): Promise<any> {
+    public async exec(throwIfMissing?: boolean): Promise<any> {
         if (throwIfMissing && this.op !== 'findOne') {
             throw newRxError('QU9', {
                 collection: this.collection.name,
@@ -304,25 +291,14 @@ export class RxQueryBase<
             });
         }
 
-
         /**
          * run _ensureEqual() here,
          * this will make sure that errors in the query which throw inside of the RxStorage,
          * will be thrown at this execution context and not in the background.
          */
-        return _ensureEqual(this as any)
-            .then(() => firstValueFrom(this.$))
-            .then(result => {
-                if (!result && throwIfMissing) {
-                    throw newRxError('QU10', {
-                        collection: this.collection.name,
-                        query: this.mangoQuery,
-                        op: this.op
-                    });
-                } else {
-                    return result;
-                }
-            });
+        await _ensureEqual(this as any);
+        const useResult = ensureNotFalsy(this._result);
+        return useResult.getValue(throwIfMissing);
     }
 
 
@@ -485,6 +461,7 @@ export class RxQueryBase<
     }
 }
 
+
 export function _getDefaultQuery<RxDocType>(): MangoQuery<RxDocType> {
     return {
         selector: {}
@@ -542,13 +519,17 @@ function _isResultsInSync(rxQuery: RxQueryBase<any, any>): boolean {
  * to ensure it does not run in parallel
  * @return true if has changed, false if not
  */
-function _ensureEqual(rxQuery: RxQueryBase<any, any>): Promise<boolean> {
+async function _ensureEqual(rxQuery: RxQueryBase<any, any>): Promise<boolean> {
+    if (rxQuery.collection.awaitBeforeReads.size > 0) {
+        await Promise.all(Array.from(rxQuery.collection.awaitBeforeReads).map(fn => fn()));
+    }
+
     // Optimisation shortcut
     if (
         rxQuery.collection.database.destroyed ||
         _isResultsInSync(rxQuery)
     ) {
-        return PROMISE_RESOLVE_FALSE;
+        return false;
     }
 
     rxQuery._ensureEqualQueue = rxQuery._ensureEqualQueue